refactor: renamed B2B_express_checkout && added aliases to mpesaclient
