+++ tests/integration/b2b_express_checkout/test_b2b_express_checkout_e2e.py
@@ -13,7 +13,7 @@ from mpesakit.auth import TokenManager
 from mpesakit.http_client import MpesaHttpClient
 import uuid
 
-from mpesakit.B2B_express_checkout import (
+from mpesakit.b2b_express_checkout import (
     B2BExpressCheckout,
     B2BExpressCheckoutRequest,
 )
